# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Kala Kendra Sweden API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: admissions
    description: Admission applications
  - name: students
    description: Enrolled students
  - name: batches
    description: Course batches
  - name: enquiries
    description: Public contact enquiries
  - name: stats
    description: Dashboard statistics
  - name: settings
    description: School settings

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # ─── ADMISSIONS ───────────────────────────────────────────────────────────
  /admissions:
    get:
      operationId: listAdmissions
      tags: [admissions]
      summary: List all admission applications
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, under_review, accepted, waitlisted, rejected]
        - name: batch
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of admissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Admission"
    post:
      operationId: createAdmission
      tags: [admissions]
      summary: Submit a new admission application (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdmissionBody"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admission"

  /admissions/{id}:
    get:
      operationId: getAdmission
      tags: [admissions]
      summary: Get a single admission application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Admission found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admission"
        "404":
          description: Not found
    patch:
      operationId: updateAdmission
      tags: [admissions]
      summary: Update admission status or notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdmissionBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admission"
    delete:
      operationId: deleteAdmission
      tags: [admissions]
      summary: Delete an admission application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted

  # ─── STUDENTS ─────────────────────────────────────────────────────────────
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all enrolled students
      parameters:
        - name: batchId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, withdrawn]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Enrol a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Student enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"

  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a single student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Student found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Remove a student record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted

  # ─── BATCHES ──────────────────────────────────────────────────────────────
  /batches:
    get:
      operationId: listBatches
      tags: [batches]
      summary: List all course batches
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of batches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Batch"
    post:
      operationId: createBatch
      tags: [batches]
      summary: Create a new batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBatchBody"
      responses:
        "201":
          description: Batch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"

  /batches/{id}:
    patch:
      operationId: updateBatch
      tags: [batches]
      summary: Update a batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBatchBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
    delete:
      operationId: deleteBatch
      tags: [batches]
      summary: Delete a batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted

  # ─── ENQUIRIES ────────────────────────────────────────────────────────────
  /enquiries:
    get:
      operationId: listEnquiries
      tags: [enquiries]
      summary: List all contact enquiries
      parameters:
        - name: read
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of enquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enquiry"
    post:
      operationId: createEnquiry
      tags: [enquiries]
      summary: Submit a contact enquiry (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnquiryBody"
      responses:
        "201":
          description: Enquiry received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"

  /enquiries/{id}:
    patch:
      operationId: updateEnquiry
      tags: [enquiries]
      summary: Mark enquiry as read or add admin notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEnquiryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"

  # ─── SETTINGS ─────────────────────────────────────────────────────────────
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get school settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolSettings"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update school settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolSettings"

  # ─── STATS ────────────────────────────────────────────────────────────────
  /stats/dashboard:
    get:
      operationId: getDashboardStats
      tags: [stats]
      summary: Aggregated admin dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    # ─── ADMISSIONS ─────────────────────────────────────────────────────────
    Admission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        applicantType:
          type: string
          enum: [adult, child]
        studentName:
          type: string
        studentDob:
          type: string
          format: date
        studentGender:
          type: string
          nullable: true
        studentEmail:
          type: string
          nullable: true
        studentPhone:
          type: string
          nullable: true
        parentName:
          type: string
          nullable: true
        parentRelationship:
          type: string
          nullable: true
        parentEmail:
          type: string
          nullable: true
        parentPhone:
          type: string
          nullable: true
        emergencyName:
          type: string
          nullable: true
        emergencyPhone:
          type: string
          nullable: true
        addressStreet:
          type: string
        addressPostal:
          type: string
        addressCity:
          type: string
        batch:
          type: string
        experience:
          type: string
          enum: [none, some, significant]
        experienceDetails:
          type: string
          nullable: true
        joiningDate:
          type: string
          format: date
          nullable: true
        medicalNotes:
          type: string
          nullable: true
        willStagePerform:
          type: string
          enum: [yes, no, maybe]
        motivation:
          type: string
          nullable: true
        referralSource:
          type: string
          nullable: true
        photoConsent:
          type: string
          enum: [yes-all, yes-internal, no]
        rulesConsent:
          type: string
          enum: [agree, disagree]
        suggestions:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, under_review, accepted, waitlisted, rejected]
        adminNotes:
          type: string
          nullable: true
        submittedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - applicantType
        - studentName
        - studentDob
        - addressStreet
        - addressPostal
        - addressCity
        - batch
        - experience
        - willStagePerform
        - photoConsent
        - rulesConsent
        - status
        - submittedAt
        - createdAt
        - updatedAt

    CreateAdmissionBody:
      type: object
      properties:
        applicantType:
          type: string
          enum: [adult, child]
        studentName:
          type: string
        studentDob:
          type: string
          format: date
        studentGender:
          type: string
        studentEmail:
          type: string
        studentPhone:
          type: string
        parentName:
          type: string
        parentRelationship:
          type: string
        parentEmail:
          type: string
        parentPhone:
          type: string
        emergencyName:
          type: string
        emergencyPhone:
          type: string
        addressStreet:
          type: string
        addressPostal:
          type: string
        addressCity:
          type: string
        batch:
          type: string
        experience:
          type: string
          enum: [none, some, significant]
        experienceDetails:
          type: string
        joiningDate:
          type: string
          format: date
        medicalNotes:
          type: string
        willStagePerform:
          type: string
          enum: [yes, no, maybe]
        motivation:
          type: string
        referralSource:
          type: string
        photoConsent:
          type: string
          enum: [yes-all, yes-internal, no]
        rulesConsent:
          type: string
          enum: [agree, disagree]
        suggestions:
          type: string
      required:
        - applicantType
        - studentName
        - studentDob
        - addressStreet
        - addressPostal
        - addressCity
        - batch
        - experience
        - willStagePerform
        - photoConsent
        - rulesConsent

    UpdateAdmissionBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, under_review, accepted, waitlisted, rejected]
        adminNotes:
          type: string

    # ─── STUDENTS ───────────────────────────────────────────────────────────
    Student:
      type: object
      properties:
        id:
          type: string
          format: uuid
        admissionId:
          type: string
          format: uuid
          nullable: true
        fullName:
          type: string
        dob:
          type: string
          format: date
          nullable: true
        batchId:
          type: string
          format: uuid
          nullable: true
        batchName:
          type: string
          nullable: true
        primaryContactName:
          type: string
          nullable: true
        primaryContactEmail:
          type: string
          nullable: true
        primaryContactPhone:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive, withdrawn]
        enrolledAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - status
        - enrolledAt
        - createdAt
        - updatedAt

    CreateStudentBody:
      type: object
      properties:
        admissionId:
          type: string
          format: uuid
        fullName:
          type: string
        dob:
          type: string
          format: date
        batchId:
          type: string
          format: uuid
        primaryContactName:
          type: string
        primaryContactEmail:
          type: string
        primaryContactPhone:
          type: string
      required: [fullName]

    UpdateStudentBody:
      type: object
      properties:
        fullName:
          type: string
        dob:
          type: string
          format: date
        batchId:
          type: string
          format: uuid
          nullable: true
        primaryContactName:
          type: string
        primaryContactEmail:
          type: string
        primaryContactPhone:
          type: string
        status:
          type: string
          enum: [active, inactive, withdrawn]

    # ─── BATCHES ────────────────────────────────────────────────────────────
    Batch:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        ageRange:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        active:
          type: boolean
        displayOrder:
          type: integer
        studentCount:
          type: integer
      required: [id, code, name, active, displayOrder, studentCount]

    CreateBatchBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        ageRange:
          type: string
        description:
          type: string
        active:
          type: boolean
        displayOrder:
          type: integer
      required: [code, name]

    UpdateBatchBody:
      type: object
      properties:
        name:
          type: string
        ageRange:
          type: string
        description:
          type: string
        active:
          type: boolean
        displayOrder:
          type: integer

    # ─── ENQUIRIES ──────────────────────────────────────────────────────────
    Enquiry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        adminNotes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, subject, message, isRead, createdAt]

    CreateEnquiryBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string
      required: [name, email, subject, message]

    UpdateEnquiryBody:
      type: object
      properties:
        isRead:
          type: boolean
        adminNotes:
          type: string

    # ─── SETTINGS ───────────────────────────────────────────────────────────
    SchoolSettings:
      type: object
      properties:
        id:
          type: integer
        monthlyFeeSek:
          type: integer
        schoolName:
          type: string
        contactEmail:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        addressLine:
          type: string
          nullable: true
        acceptingApplications:
          type: boolean
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - monthlyFeeSek
        - schoolName
        - acceptingApplications
        - updatedAt

    UpdateSettingsBody:
      type: object
      properties:
        monthlyFeeSek:
          type: integer
        schoolName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        addressLine:
          type: string
        acceptingApplications:
          type: boolean

    # ─── STATS ──────────────────────────────────────────────────────────────
    DashboardStats:
      type: object
      properties:
        totalStudents:
          type: integer
        activeStudents:
          type: integer
        totalBatches:
          type: integer
        activeBatches:
          type: integer
        pendingAdmissions:
          type: integer
        underReviewAdmissions:
          type: integer
        totalAdmissions:
          type: integer
        unreadEnquiries:
          type: integer
        recentAdmissions:
          type: array
          items:
            $ref: "#/components/schemas/Admission"
      required:
        - totalStudents
        - activeStudents
        - totalBatches
        - activeBatches
        - pendingAdmissions
        - underReviewAdmissions
        - totalAdmissions
        - unreadEnquiries
        - recentAdmissions